advisory-chat-service/requirements.txt
django>=4.2
djangorestframework>=3.14
requests>=2.31
pika>=1.3
psycopg2-binary>=2.9
sentence-transformers>=2.2
pgvector>=0.3
openai>=1.0
